#139 - change date is retrieved from server and set in entities on push

+++ sormas-app/app/src/main/java/de/symeda/sormas/app/caze/SyncCasesTask.java
@@ -11,15 +11,11 @@ import com.j256.ormlite.logger.LoggerFactory;
 import java.util.List;
 
 import de.symeda.sormas.api.caze.CaseDataDto;
-import de.symeda.sormas.api.caze.PreviousHospitalizationDto;
-import de.symeda.sormas.api.sample.SampleDto;
 import de.symeda.sormas.app.backend.caze.CaseDtoHelper;
 import de.symeda.sormas.app.backend.common.AdoDtoHelper;
 import de.symeda.sormas.app.backend.common.AdoDtoHelper.DtoGetInterface;
 import de.symeda.sormas.app.backend.common.DatabaseHelper;
 import de.symeda.sormas.app.backend.config.ConfigProvider;
-import de.symeda.sormas.app.backend.hospitalization.PreviousHospitalizationDtoHelper;
-import de.symeda.sormas.app.backend.sample.SampleDtoHelper;
 import de.symeda.sormas.app.backend.user.User;
 import de.symeda.sormas.app.person.SyncPersonsTask;
 import de.symeda.sormas.app.rest.RetroProvider;
@@ -55,7 +51,7 @@ public class SyncCasesTask extends AsyncTask<Void, Void, Void> {
 
         new CaseDtoHelper().pushEntities(new AdoDtoHelper.DtoPostInterface<CaseDataDto>() {
             @Override
-            public Call<Integer> postAll(List<CaseDataDto> dtos) {
+            public Call<Long> postAll(List<CaseDataDto> dtos) {
                 return RetroProvider.getCaseFacade().postAll(dtos);
             }
         }, DatabaseHelper.getCaseDao());
