+++ sormas-app/app/src/main/java/de/symeda/sormas/app/event/SyncEventParticipantsTask.java
@@ -16,7 +16,6 @@ import de.symeda.sormas.app.backend.event.EventParticipantDtoHelper;
 import de.symeda.sormas.app.backend.user.User;
 import de.symeda.sormas.app.rest.RetroProvider;
 import de.symeda.sormas.app.util.Callback;
-import de.symeda.sormas.app.visit.SyncVisitsTask;
 import retrofit2.Call;
 
 /**
@@ -45,7 +44,7 @@ public class SyncEventParticipantsTask extends AsyncTask<Void, Void, Void> {
 
         new EventParticipantDtoHelper().pushEntities(new DtoPostInterface<EventParticipantDto>() {
             @Override
-            public Call<Integer> postAll(List<EventParticipantDto> dtos) {
+            public Call<Long> postAll(List<EventParticipantDto> dtos) {
                 // TODO postAll should return the date&time the server used as modifiedDate
                 return RetroProvider.getEventParticipantFacade().postAll(dtos);
             }
