+++ sormas-app/app/src/main/java/de/symeda/sormas/app/rest/EventFacadeRetro.java
@@ -15,5 +15,5 @@ public interface EventFacadeRetro {
     Call<List<EventDto>> getAll(@Path("user") String userUuid, @Path("since") long since);
 
     @POST("events/push")
-    Call<Integer> postAll(@Body List<EventDto> dtos);
+    Call<Long> postAll(@Body List<EventDto> dtos);
 }
